move logging tests into specific file
